requirements.txt
absl-py==1.2.0
attrs==22.1.0
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
fonttools==4.35.0
idna==3.3
imutils==0.5.4
kiwisolver==1.4.4
matplotlib==3.5.3
mediapipe==0.8.10.1
numpy==1.23.2
opencv-contrib-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.20.1
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
tk==0.1.0
toml==0.10.2
urllib3==1.26.11
